\name{neptune-package}
\alias{neptune-package}
\alias{neptune}
\docType{package}
\title{
\packageTitle{neptune}
}
\description{
\packageDescription{neptune}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{neptune}
\packageIndices{neptune}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{neptune}

Maintainer: \packageMaintainer{neptune}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
